+++ src/main/java/insane96mcp/survivalreimagined/module/mining/forging/ForgeRecipeBookComponent.java
@@ -76,8 +76,6 @@ public class ForgeRecipeBookComponent extends RecipeBookComponent {
 
     }
 
-
-
     public void renderGhostRecipeAmount(GuiGraphics guiGraphics, int pLeftPos, int pTopPos, boolean p_100326_, float pPartialTick) {
         this.ghostRecipeAmount.render(guiGraphics, this.minecraft, pLeftPos, pTopPos, p_100326_, pPartialTick);
     }