+++ src/main/java/insane96mcp/survivalreimagined/module/items/itemstats/ItemStats.java
@@ -21,6 +21,8 @@ import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.Attribute;
@@ -33,6 +35,7 @@ import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.event.ItemAttributeModifierEvent;
+import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
@@ -316,7 +319,29 @@ public class ItemStats extends JsonFeature {
 	}
 
 	public static boolean isBroken(ItemStack stack) {
-		return shouldNotBreak(stack) && stack.getDamageValue() >= stack.getMaxDamage() - 1;
+		return stack.isDamageableItem() && shouldNotBreak(stack) && stack.getDamageValue() >= stack.getMaxDamage() - 1;
+	}
+
+	@SubscribeEvent
+	public void onPlayerTick(TickEvent.PlayerTickEvent event) {
+		if (!this.isEnabled()
+				|| !unbreakableItems
+				|| event.player.level().isClientSide
+				|| event.phase == TickEvent.Phase.START
+				|| event.player.tickCount % 20 != event.player.getId() % 20)
+			return;
+
+		for (ItemStack stack : event.player.getArmorSlots()) {
+			if (stack.isEmpty() || !isBroken(stack))
+				continue;
+			event.player.level().playSound(null, event.player, SoundEvents.ALLAY_HURT, SoundSource.PLAYERS, 0.7f, 2f);
+			EquipmentSlot equipmentSlot = Player.getEquipmentSlotForItem(stack);
+			if (stack.getItem() instanceof Equipable) {
+				event.player.setItemSlot(equipmentSlot, ItemStack.EMPTY);
+				if (!event.player.addItem(stack))
+					event.player.drop(stack, true);
+			}
+		}
 	}
 
 	@SubscribeEvent
@@ -342,21 +367,6 @@ public class ItemStats extends JsonFeature {
 
 	@SubscribeEvent
 	public void onBlockRightClick(PlayerInteractEvent.RightClickBlock event) {
-		if (!this.isEnabled()
-				|| !unbreakableItems)
-			return;
-
-		ItemStack stack = event.getItemStack();
-		if (stack.getMaxDamage() == 0)
-			return;
-		if (isBroken(stack)) {
-			event.setCanceled(true);
-			event.getEntity().displayClientMessage(Component.translatable(BROKEN_ITEM_LANG), true);
-		}
-	}
-
-	@SubscribeEvent
-	public void onItemUse(PlayerInteractEvent.RightClickItem event) {
 		if (!this.isEnabled()
 				|| !unbreakableItems)
 			return;