+++ changelog.md
@@ -5,6 +5,7 @@
   * Added options to change air consumed and regained when under/out of water
 * Reduced stamina consumed when swimming
 * Using recipe book to place Forge recipes now works
+* You can no longer eat when full health
 * Fixed iron shields not repairable
 
 ## 0.17.2