+++ changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## Upcoming
-* You now regain air at the same rate at which is consumed (so much slower)
+* You now regain air at the same rate at which is consumed (so, much slower)
   * Added options to change air consumed and regained when under/out of water
+* Equipping broken armor now falls off
 * Reduced stamina consumed when swimming
 * Using recipe book to place Forge recipes now works
 * You can no longer eat when full health